refactor(firestore-bigquery-export): importer script code
